+++ cmd/dno/main_test.go
@@ -1,0 +1,7 @@
+package main
+
+import "testing"
+
+func FakeTest(t *testing.T) {
+	// todo
+}
