--- main_test.go
@@ -1,7 +1,0 @@
-package main
-
-import "testing"
-
-func FakeTest(t *testing.T) {
-	// todo
-}
